Añadidos cambios de feat/scaffold-spring-boot-app
Se han separado en distintos archivos cada clase de la aplicación Spring
Ignorar Gradle.properties
Versión 16 de java (cambia dockerfile y guild.gradle)
