--- gradle.properties
@@ -1,2 +1,0 @@
-gpr.key=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
-gpr.user=irm@us.es
