+++ src/main/java/us/mitfs/samples/auditserver/MetricController.java
@@ -1,9 +1,10 @@
 /**
  * 
  */
-package us.mitfs.samples.server;
+package us.mitfs.samples.auditserver;
 
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 public class MetricController {
+	
 		
 		@GetMapping("/metric")
 		public MetricInfo metric(@RequestParam(value = "name", defaultValue = "all") String name) {
