+++ src/main/java/us/mitfs/samples/auditserver/ServletInitializer.java
@@ -1,13 +1,15 @@
-package us.mitfs.samples.server;
+package us.mitfs.samples.auditserver;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+import us.mitfs.samples.auditserver.AuditserverApplication;
 
 public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(ServerApplication.class);
+		return application.sources(AuditserverApplication.class);
 	}
 
 }
