--- src/main/java/us/mitfs/samples/server/ServerApplication.java
@@ -1,13 +1,0 @@
-package us.mitfs.samples.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
-	}
-
-}
